Big unit testing rework
