+++ CanDI/candi/__init__.py
@@ -1,0 +1,4 @@
+from CanDI.candi import data
+data = data.Data() #Global object data instantiated on import required for access by GeneQuery Objects
+from CanDI.candi.candi import (Gene, CellLine, Organelle, Cancer, CellLineCluster, GeneCluster)
+
