+++ CanDI/setup/manager.py
