+++ scripts/install.py
@@ -1,6 +1,7 @@
-from manager import Manager
+from CanDI.setup.manager import Manager
 
 if __name__ == "__main__":
+
     m = Manager()
     m.get_depmap_info()
     m.write_config(m.cfig_path, m.parser)
