+++ scripts/reset_config.py
